requirements.txt
discord.py[voice]==1.7.3
youtube_dl==2021.6.6
python-dotenv==0.19.0
lyricsgenius==3.0.1